refactor: httpcache
