--- httpcache/doc.go
@@ -1,2 +1,0 @@
-// Package httpcache provides net/http middleware backed by Kioshun.
-package httpcache
