+++ httpcache/pattern.go
