+++ httpcache/pattern_test.go
